% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostriches.R
\docType{data}
\name{ostriches}
\alias{ostriches}
\title{Height and mass of four ostriches}
\format{
A tibble with 4 rows and 3 variables:
\describe{
  \item{name}{a character denoting ostrich name}
  \item{height_m}{a number denoting ostrich height in meters}
  \item{mass_kg}{a number denoting ostrich mass in kilograms}
  ...
}
}
\source{
\url{http://www.ucsb.edu/}
}
\usage{
ostriches
}
\description{
A toy dataset with ostrich size data.
}
\keyword{datasets}
